fix: sending the API request to the Provider
