--- src/test/renderWithRouter.js
@@ -1,13 +1,0 @@
-// import React from 'react'
-// import { render } from '@testing-library/react'
-// import { Router } from 'react-router-dom'
-// import { createMemoryHistory } from 'history'
-
-// const renderWithRouter = (component) => {
-//   const history = createMemoryHistory()
-//   return ({
-//     ...render(<Router history={history}>{component}</Router>), history,
-//   })
-// }
-
-// export default renderWithRouter
